Commit Base Files
Dockerfile that uses apt-get to install the Clang toochain ontop of the
CppUTest with GCC image that I already made.  I mostly use this for the
    scan-build tool around my GCC based unit test builds.

Copy CONTRIBUTING.md & README.md over from my tafthorne/cpputest-gcc
project and update to add Clang instead of cppucheck.

I chose to base the image on my cpputest-gcc one as I expect CppUTest
will need to be updated without using a pre-packaged version, if they
version I have proves to old.  I know that the Debian version is quite a
long way behind the latest as it is.  This re-use should allow me to
handle adding a new version of cpputest once in the parent image.

I have already published the tafthorne/cpputest-gcc Dockerfile on GitHub
and setup Docker Hub to automatically generate it.  It is available for
public use and inspection.